Extract OAuthPolicyData to a top-level file with a from() factory method
  Move OAuthPolicyData out of OAuthUtils into its own file so integrations
  can reference it directly, and add a static from(AuthenticationPolicyUnion)
  convenience factory that delegates to OAuthUtils.from().
